A Python utility for visualizing Time-Dependent Density Functional Theory (TDDFT) output from Orca, generating absorption spectra and comprehensive data summaries
This repository contains a Python script tailored for chemists and researchers working with Time-Dependent Density Functional Theory (TDDFT) outputs from Orca. The script offers essential tools for spectral analysis and data handling:

Spectra Visualization: It processes TDDFT output files to generate electronic absorption spectra, providing options to visualize the data using Gaussian or Lorentzian curves. This feature aids in the detailed examination of spectral characteristics derived from TDDFT calculations.

Experimental Data Integration: The script allows for the integration of experimental data (expected in a file named Experiment.txt) to directly compare with the theoretical spectra. In cases where experimental data is not available, it utilizes a mock dataset to demonstrate the comparative analysis.

Output Summary: A summary of the TDDFT data is compiled into a DOCX file. This summary includes essential data points such as energy in cm-1, nm, and eV, offering a convenient and accessible presentation of the results.

Graphical Results: Shows all the contributions from all electronic states as well!

This tool aims to simplify the process of analyzing TDDFT data by providing a straightforward and user-friendly approach to visualizing and comparing spectral data. It's particularly useful for quick assessments for Orca software.